+++ app/src/main/java/com/yuechen/whiteboard/Model/TodoItem.java
@@ -3,7 +3,7 @@ package com.yuechen.whiteboard.Model;
 public class TodoItem {
     public long _id = -1;
     public String title;
-    public int folderID;
+    public long folderID;
     public String endDateTime;
     public String note;
     public boolean finished;
