+++ app/src/main/java/com/yuechen/whiteboard/adapter/FolderAdapter.java
@@ -59,7 +59,7 @@ public class FolderAdapter extends RecyclerView.Adapter<FolderAdapter.ViewHolder
             } else {
                 Intent intent = new Intent(context, TodoListActivity.class);
                 Bundle bundle = new Bundle();
-                bundle.putInt("id", folder.getFolderId());
+                bundle.putLong("id", folder.getFolderId());
                 bundle.putBoolean("isLesson", false);
                 intent.putExtras(bundle);
                 context.startActivity(intent);
